+++ src/main/java/de/samply/exporter/ExporterConst.java
@@ -58,6 +58,14 @@ public class ExporterConst {
     public final static String XML_FILE_MERGER_ROOT_ELEMENT = "XML_FILE_MERGER_ROOT_ELEMENT";
     public final static String CSV_SEPARATOR_REPLACEMENT = "CSV_SEPARATOR_REPLACEMENT";
     public final static String SITE = "SITE";
+    public static final String WEBCLIENT_REQUEST_TIMEOUT_IN_SECONDS = "WEBCLIENT_REQUEST_TIMEOUT_IN_SECONDS";
+    public static final String WEBCLIENT_CONNECTION_TIMEOUT_IN_SECONDS = "WEBCLIENT_CONNECTION_TIMEOUT_IN_SECONDS";
+    public static final String WEBCLIENT_TCP_KEEP_IDLE_IN_SECONDS = "WEBCLIENT_TCP_KEEP_IDLE_IN_SECONDS";
+    public static final String WEBCLIENT_TCP_KEEP_INTERVAL_IN_SECONDS = "WEBCLIENT_TCP_KEEP_INTERVAL_IN_SECONDS";
+    public static final String WEBCLIENT_TCP_KEEP_CONNECTION_NUMBER_OF_TRIES = "WEBCLIENT_TCP_KEEP_CONNECTION_NUMBER_OF_TRIES";
+    public static final String WEBCLIENT_MAX_NUMBER_OF_RETRIES = "WEBCLIENT_MAX_NUMBER_OF_RETRIES";
+    public static final String WEBCLIENT_TIME_IN_SECONDS_AFTER_RETRY_WITH_FAILURE = "WEBCLIENT_TIME_IN_SECONDS_AFTER_RETRY_WITH_FAILURE";
+    public static final String WEBCLIENT_BUFFER_SIZE_IN_BYTES = "WEBCLIENT_BUFFER_SIZE_IN_BYTES";
 
     // Spring Values (SV)
     public final static String HEAD_SV = "${";
@@ -108,7 +116,22 @@ public class ExporterConst {
     public final static String CSV_SEPARATOR_REPLACEMENT_SV =
             HEAD_SV + CSV_SEPARATOR_REPLACEMENT + ": " + BOTTOM_SV;
     public final static String SITE_SV = HEAD_SV + SITE + ":" + DEFAULT_SITE + BOTTOM_SV;
-
+    public static final String WEBCLIENT_REQUEST_TIMEOUT_IN_SECONDS_SV =
+            HEAD_SV + WEBCLIENT_REQUEST_TIMEOUT_IN_SECONDS + ":10" + BOTTOM_SV; // default 10 seconds
+    public static final String WEBCLIENT_CONNECTION_TIMEOUT_IN_SECONDS_SV =
+            HEAD_SV + WEBCLIENT_CONNECTION_TIMEOUT_IN_SECONDS + ":5" + BOTTOM_SV; // default 5 seconds
+    public static final String WEBCLIENT_TCP_KEEP_IDLE_IN_SECONDS_SV =
+            HEAD_SV + WEBCLIENT_TCP_KEEP_IDLE_IN_SECONDS + ":30" + BOTTOM_SV; // default 30 seconds
+    public static final String WEBCLIENT_TCP_KEEP_INTERVAL_IN_SECONDS_SV =
+            HEAD_SV + WEBCLIENT_TCP_KEEP_INTERVAL_IN_SECONDS + ":10" + BOTTOM_SV; // default 10 seconds
+    public static final String WEBCLIENT_TCP_KEEP_CONNECTION_NUMBER_OF_TRIES_SV =
+            HEAD_SV + WEBCLIENT_TCP_KEEP_CONNECTION_NUMBER_OF_TRIES + ":3" + BOTTOM_SV; // default 3 tries
+    public static final String WEBCLIENT_MAX_NUMBER_OF_RETRIES_SV =
+            HEAD_SV + WEBCLIENT_MAX_NUMBER_OF_RETRIES + ":2" + BOTTOM_SV; // default 2 retries
+    public static final String WEBCLIENT_TIME_IN_SECONDS_AFTER_RETRY_WITH_FAILURE_SV =
+            HEAD_SV + WEBCLIENT_TIME_IN_SECONDS_AFTER_RETRY_WITH_FAILURE + ":1" + BOTTOM_SV; // default 1 second
+    public static final String WEBCLIENT_BUFFER_SIZE_IN_BYTES_SV =
+            HEAD_SV + WEBCLIENT_BUFFER_SIZE_IN_BYTES + ":8192" + BOTTOM_SV; // default 8 KB buffer
 
     // REST Paths
     public static final String INFO = "/info";