+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.1 - 2025-04-10]
+## [1.1.1 - 2025-05-21]
 ### Added
-
+- WebClient Factory
 
 ## [1.1.0 - 2025-04-10]
 ### Added